Move heavy lifting parts into a dedicated module
That way, you should be able to reuse most of the action-code, but replace
the GitHub Actions specific part with other wrappers. For example, a simple
CLI command could be used to trigger deployments as well.
